+++ app/views/archives_requests/new.html.erb
@@ -29,15 +29,14 @@
           <% end %>
         </div>
       </div>
-      <%= form_with id: 'archives-request', model: @ead_request, url: archives_requests_path, class: 'accordion', html: { autocomplete: "off" }, data: { controller: 'accordion-form item-selector archives-request', action: 'item-selector:changed->archives-request#onSelectedItemsValueChanged item-selector:changed->accordion-form#typeValueChanged item-selector:changed->accordion-form#reenableNextButtons', turbo: false } do |f| %>
+      <%= form_with id: 'archives-request', model: @ead_request, url: archives_requests_path, class: 'accordion', html: { autocomplete: "off" }, data: { controller: 'accordion-form item-selector archives-request update-request-type', action: 'item-selector:changed->archives-request#onSelectedItemsValueChanged item-selector:changed->accordion-form#typeValueChanged item-selector:changed->accordion-form#reenableNextButtons', turbo: false } do |f| %>
         <%= f.hidden_field :ead_url %>
         <%= render AccordionStepComponent.new(id: 'request_type', step_index: step_enum.next, form_id: f.id) do |component| %>
           <% component.with_title.with_content('Request type') %>
           <% component.with_body do %>
             <fieldset class="ps-2">
               <div>
-                <!--TODO: refactor out #updateType in a new JS controller. Or pick one. This is wrong -->
-                <%= f.radio_button :request_type, 'pickup', :class => 'form-check-input', :required => true, data: { action: 'archives-request#updateRequestType' } %>
+                <%= f.radio_button :request_type, 'pickup', :class => 'form-check-input', :required => true, data: { action: 'update-request-type#updateType' } %>
                 <%= f.label :request_type_pickup, :class => 'form-check-label ms-1' do %>
                   Reading room appointment
                 <% end %>
@@ -47,7 +46,7 @@
                 </span>
               </div>
               <div class="mt-2">
-                <%= f.radio_button :request_type, 'scan', :class => 'form-check-input', data: { action: 'archives-request#updateRequestType' } %>
+                <%= f.radio_button :request_type, 'scan', :class => 'form-check-input', data: { action: 'update-request-type#updateType' } %>
                 <%= f.label :request_type_scan, :class => 'form-check-label ms-1' do %>
                   Digitization
                 <% end %>
