+++ app/views/patron_requests/_scan_or_pickup.html.erb
@@ -11,7 +11,7 @@
     </div>
   <% end %>
   <div>
-    <%= f.radio_button :request_type, 'scan', :class => 'form-check-input', :required => true, data: { action: 'patron-request#updateType' } %>
+    <%= f.radio_button :request_type, 'scan', :class => 'form-check-input', :required => true, data: { action: 'update-request-type#updateType' } %>
     <%= f.label :request_type_scan, :class => 'form-check-label ms-1' do %>
       Email digital scan
     <% end %>
@@ -25,7 +25,7 @@
     </div>
   </div>
   <div class="mt-4">
-    <%= f.radio_button :request_type, 'pickup', :class => 'form-check-input', data: { action: 'patron-request#updateType' } %>
+    <%= f.radio_button :request_type, 'pickup', :class => 'form-check-input', data: { action: 'update-request-type#updateType' } %>
     <%= f.label :request_type_pickup, :class => 'form-check-label ms-1' do %>
       Pickup physical item
     <% end %>
